+++ src/dfsi/rtp/MotStartVoiceFrame.h
@@ -1,0 +1,79 @@
+// SPDX-License-Identifier: GPL-2.0-only
+/**
+* Digital Voice Modem - Common Library
+* GPLv2 Open Source. Use is subject to license terms.
+* DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+*
+* @package DVM / DFSI peer application
+* @derivedfrom MMDVMHost (https://github.com/g4klx/MMDVMHost)
+* @license GPLv2 License (https://opensource.org/licenses/GPL-2.0)
+*
+*   Copyright (C) 2024 Patrick McDonnell, W3AXL
+*   Copyright (C) 2024 Bryan Biedenkapp, N2PLL
+*
+*/
+#if !defined(__MOT_START_VOICE_FRAME_H__)
+#define __MOT_START_VOICE_FRAME_H__
+
+#include "Defines.h"
+#include "common/Defines.h"
+#include "common/Log.h"
+#include "common/Utils.h"
+#include "rtp/RtpDefines.h"
+#include "rtp/MotStartOfStream.h"
+#include "rtp/MotFullRateVoice.h"
+
+namespace p25
+{
+    namespace dfsi
+    {
+        // ---------------------------------------------------------------------------
+        //  Class Declaration
+        //      Implements a P25 Motorola voice frame 1/10 start.
+        // 
+        // Byte 0               1               2               3
+        // Bit  0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7
+        //     +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+        //     |   Encoded Motorola Start of Stream                            |
+        //     +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+        //     |   ICW Flag ?  |     RSSI      |  RSSI Valid   |     RSSI      |
+        //     +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+        //     |   Adj MM ?    |    Full Rate Voice Frame                      |
+        //     +-+-+-+-+-+-+-+-+                                               +
+        //     |                                                               |
+        //     +                                                               +
+        //     |                                                               |
+        //     +               +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+        //     |               |
+        //     +=+=+=+=+=+=+=+=+
+        // ---------------------------------------------------------------------------
+
+        class HOST_SW_API MotStartVoiceFrame {
+        public:
+            static const uint8_t LENGTH = 22;
+
+            ICWFlag icw;
+            uint8_t rssi;
+            RssiValidityFlag rssiValidity;
+            uint8_t nRssi;
+            uint8_t adjMM;
+
+            MotStartOfStream* startOfStream;
+            MotFullRateVoice* fullRateVoice;
+
+            /// <summary>Initializes a copy instance of the MotStartVoiceFrame class.</summary>
+            MotStartVoiceFrame();
+            /// <summary>Initializes a copy instance of the MotStartVoiceFrame class.</summary>
+            MotStartVoiceFrame(uint8_t* data);
+            /// <summary>Finalizes a instance of the MotStartVoiceFrame class.</summary>
+            ~MotStartVoiceFrame();
+
+            /// <summary>Decode a start voice frame.</summary>
+            bool decode(const uint8_t* data);
+            /// <summary>Encode a start voice frame.</summary>
+            void encode(uint8_t* data);
+        };
+    } // namespace dfsi
+} // namespace p25
+
+#endif // __MOT_START_VOICE_FRAME_H__
