+++ src/dfsi/rtp/RtpFrames.h
@@ -1,0 +1,26 @@
+// SPDX-License-Identifier: GPL-2.0-only
+/**
+* Digital Voice Modem - Modem Host Software
+* GPLv2 Open Source. Use is subject to license terms.
+* DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+*
+* @package DVM / DFSI peer application
+* @derivedfrom MMDVMHost (https://github.com/g4klx/MMDVMHost)
+* @license GPLv2 License (https://opensource.org/licenses/GPL-2.0)
+*
+*   Copyright (C) 2024 Patrick McDonnell, W3AXL
+*   Copyright (C) 2024 Bryan Biedenkapp, N2PLL
+*
+*/
+#if !defined(__RTP_FRAMES_H__)
+#define __RTP_FRAMES_H__
+
+#include "Defines.h"
+
+#include "rtp/MotFullRateVoice.h"
+#include "rtp/MotStartOfStream.h"
+#include "rtp/MotStartVoiceFrame.h"
+#include "rtp/MotVoiceHeader1.h"
+#include "rtp/MotVoiceHeader2.h"
+
+#endif // __RTP_FRAMES_H__
